ObjectManagerを分離
