+++ ShaderProject/ShaderProject.vcxproj
@@ -182,6 +182,7 @@
     <ClCompile Include="DirectX.cpp" />
     <ClCompile Include="DirectXMathUtil.cpp" />
     <ClCompile Include="InclutionShot.cpp" />
+    <ClCompile Include="ObjectManager.cpp" />
     <ClCompile Include="ScenePause.cpp" />
     <ClCompile Include="SceneTitle.cpp" />
     <ClCompile Include="Util.cpp" />
@@ -225,6 +226,7 @@
     <ClInclude Include="DirectX.h" />
     <ClInclude Include="DirectXMathUtil.h" />
     <ClInclude Include="InclutionShot.h" />
+    <ClInclude Include="ObjectManager.h" />
     <ClInclude Include="ScenePause.h" />
     <ClInclude Include="SceneTitle.h" />
     <ClInclude Include="Util.h" />
